refactor: remove AvmVerificationKeyData
