+++ yarn-project/bb-prover/src/avm_proving.test.ts
@@ -1,10 +1,10 @@
 import {
   AvmCircuitInputs,
-  AvmVerificationKeyData,
   FunctionSelector,
   Gas,
   GlobalVariables,
   SerializableContractInstance,
+  VerificationKeyData,
 } from '@aztec/circuits.js';
 import { Fr } from '@aztec/foundation/fields';
 import { createDebugLogger } from '@aztec/foundation/log';
@@ -131,7 +131,7 @@ const proveAndVerifyAvmTestContract = async (
   // Then we test VK extraction and serialization.
   const succeededRes = proofRes as BBSuccess;
   const vkData = await extractAvmVkData(succeededRes.vkPath!);
-  AvmVerificationKeyData.fromBuffer(vkData.toBuffer());
+  VerificationKeyData.fromBuffer(vkData.toBuffer());
 
   // Then we verify.
   const rawVkPath = path.join(succeededRes.vkPath!, 'vk');
